rework dirty writeout code
The basic problem is that we can't block dirty writeout by
marking an object dirty.  So, we have to make a way to
remember dirty objects in a non-blocking way.

So, we create a non-size-limited Queue for these, then tie
that queue back in to the existing one that _can_ block.